feat: Move manifest file to cozy-client models
BREAKING CHANGE: If you were importing manifest from
'cozy-client/dist/manifest` you have to either import it from the root
 or from the models.
