+++ packages/cozy-client/src/models/manifest.spec.js
