+++ packages/cozy-client/types/models/manifest.d.ts
